refactor: separate compute functions
